Add /api resource to APIG
This is to use the APIG with CloudFront because we will call CloudFront
with /api path to call the APIG behind it.

Also note that the old APIG is not deleted yet and its values are
explicitly exported with exportValue method. This is to avoid deployment
error as they are still referred by CloudFront stack.